cpp-port: batch 78 - batch-port 53 TRIVIAL hw/ files to C++
Used batch-port.py to rename 189 TRIVIAL hw/ files, then reverted
121 that failed to compile (header blockers, nested designators,
compound literals, linker errors from missing extern "C").

Manually fixed:
- hw/audio/fmopl.cpp: narrowing double-to-uint32_t in KSL_TABLE and
  MUL_TABLE (added KSL/MUL cast macros), void* malloc casts,
  sign-compare fix
- hw/audio/gusemu_hal.cpp: const char cast -> static_cast
- hw/char/serial-isa.cpp: uint32_t vs -1 sign-compare fixes
- hw/i386/kvm/xen-stubs.cpp: added extern "C" for function defs
- hw/i386/kvm/xen_evtchn.h: added extern "C" guards
- hw/i386/kvm/xen_primary_console.h: added extern "C" guards
- include/hw/virtio/vhost-backend.h: renamed 3 conflicting typedefs
  to *_op suffix, added extern "C" guards

+++ hw/char/serial-isa.cpp
@@ -60,7 +60,7 @@ static void serial_isa_realizefn(DeviceState *dev, Error **errp)
     ISASerialState *isa = ISA_SERIAL(dev);
     SerialState *s = &isa->state;
 
-    if (isa->index == -1) {
+    if (isa->index == static_cast<uint32_t>(-1)) {
         isa->index = index;
     }
     if (isa->index >= MAX_ISA_SERIAL_PORTS) {
@@ -68,10 +68,10 @@ static void serial_isa_realizefn(DeviceState *dev, Error **errp)
                    MAX_ISA_SERIAL_PORTS);
         return;
     }
-    if (isa->iobase == -1) {
+    if (isa->iobase == static_cast<uint32_t>(-1)) {
         isa->iobase = isa_serial_io[isa->index];
     }
-    if (isa->isairq == -1) {
+    if (isa->isairq == static_cast<uint32_t>(-1)) {
         isa->isairq = isa_serial_irq[isa->index];
     }
     index++;
